+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.2.1 / Unreleased
 
 * [FEATURE] Source can now be a plain binary file. The file will be simply fetched and placed in the container verbatim.
+* [FIX] Handle dependencies with alternatives during build install